Change migration generator to reversible style
The [primary] way of migrating in ActiveRecord is using the new reversible
migrations with the `change` method. The generator was using the old
`up/down` style, so this commit changes that. It *should* be more common
to use the `change` style than to use the `up/down` style, so this
change should generate better migration files.

[primary]:
http://guides.rubyonrails.org/migrations.html#using-the-change-method