Switch to ConnectionUrlResolver
Why you made the change:

I did this because ActiveRecord::Base.connection.pool.spec.config was
trying to establish a connection and that was causing problems when
trying to run rake db:create because the database didn't already exist.
I believe this only happens when using the DATABASE_URL environment
variable instead of the config/database.yml.